feat(api): Normalizes registerAPI
This PR removes the `onFoundTarget` action in favor of more general
`registerAPI`. This follows the pattern established by libraries like
ember-power-select and ember-basic-dropdown, and the pattern that
vertical-collection will be using for its imperative API in the near
future (specifically for searching in a collection manually).

This change also gives users an imperative way to trigger changes like
`update` and `enableEventListeners`. Parent components will often have
the context necessary to trigger these, so it makes sense for them to be
able to do so manually when needed. Finally, it adds `popperElement` to
the API, which is necessary for parent components that want to track the
element (for animations for instance) but otherwise would have no way to
since it is moved in the DOM.