refactor(raf): Remove requestAnimationFrame
This PR began with an attempt to replace RAF with the new
ember-raf-scheduler, but after testing I realized that our usage of
RAF doesn't appear to actually have any performance benefits. With that
in mind, it definitely makes more sense to not complicate the runloop
and remove it entirely for now.