fix: use @ember/render-modifiers instead of ember-ref-modifier to avoid timing bugs
The way the {{ref}} modifier defers its sets causes unpredictable timing issues in popovers. Using the simpler {{did-insert}} and {{will-destroy}} modifiers is more robust and fixes the timing issues I've encountered. I expect it also fixes kybishop/ember-popper#126. Doing it this way also avoids needing to hook into the set() on _popperElement, because with {{did-insert}} we get notified directly when the element is inserted.